requirements.txt
torch
torchvision
numpy
pillow
scikit-learn
matplotlib